Created "mix solution" task to get problem solutions in one command
In order to simplify the process of obtaining the solution of a specific
exercise, we've created a mix task that will run the tests and, if they
pass, call the "solution" method for the module corresponding a specific
day.

Example: `mix solution 2` will return the solution for day 2.

The command still can undergo some improvements (obtain both solutions
instead of just the one for part 1, run only the tests belonging to that
module...) but it's a good starting point